tree-construction/tests3.dat

#data
<head></head><style></style>
#errors
No DOCTYPE
<style> in after-head mode
#document
| <html>
|   <head>
|     <style>
|   <body>

#data
<head></head><script></script>
#errors
No DOCTYPE
<script> in after-head mode
#document
| <html>
|   <head>
|     <script>
|   <body>

#data
<head></head><!-- --><style></style><!-- --><script></script>
#errors
No DOCTYPE
<style> in after-head mode
#document
| <html>
|   <head>
|     <style>
|     <script>
|   <!--   -->
|   <!--   -->
|   <body>

#data
<head></head><!-- -->x<style></style><!-- --><script></script>
#errors
No DOCTYPE
#document
| <html>
|   <head>
|   <!--   -->
|   <body>
|     "x"
|     <style>
|     <!--   -->
|     <script>

#data
<!DOCTYPE htML><html><head></head><body><pre>
</pre></body></html>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <pre>

#data
<!DOCTYPE htML><html><head></head><body><pre>

foo</pre></body></html>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <pre>
|       "foo"


#data
<!DOCTYPE htML><html><head></head><body><pre>

foo
</pre></body></html>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <pre>
|       "foo
"

#data
<!DOCTYPE htML><html><head></head><body><pre>x</pre><span>
</span></body></html>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <pre>
|       "x"
|     <span>
|       "
"

#data
<!DOCTYPE htML><html><head></head><body><pre>x
y</pre></body></html>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <pre>
|       "x
y"

#data
<!DOCTYPE htML><html><head></head><body><pre>x<div>
y</pre></body></html>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <pre>
|       "x"
|       <div>
|         "
| y"

#data
<!DOCTYPE htML><HTML><META><HEAD></HEAD></HTML>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|     <meta>
|   <body>

#data
<!DOCTYPE htML><HTML><HEAD><head></HEAD></HTML>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>

#data
<textarea>foo<span>bar</span><i>baz
#errors
#document
| <html>
|   <head>
|   <body>
|     <textarea>
|       "foo<span>bar</span><i>baz"

#data
<title>foo<span>bar</em><i>baz
#errors
#document
| <html>
|   <head>
|     <title>
|       "foo<span>bar</em><i>baz"
|   <body>

#data
<!DOCTYPE htML><textarea>
</textarea>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <textarea>

#data
<!DOCTYPE htML><textarea>

foo</textarea>
#errors
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <textarea>
|       "foo"

#data
<!DOCTYPE htML><html><head></head><body><ul><li><div><p><li></ul></body></html>
#errors
Missing end tag (div)
#document
| <!DOCTYPE htML>
| <html>
|   <head>
|   <body>
|     <ul>
|       <li>
|         <div>
|           <p>
|       <li>

#data
<!doctype html><nobr><nobr><nobr>
#errors
Unexpected end of file.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <nobr>
|     <nobr>
|     <nobr>

#data
<!doctype html><nobr><nobr></nobr><nobr>
#errors
Unexpected end of file.
#document
| <!DOCTYPE html>
| <html>
|   <head>
|   <body>
|     <nobr>
|     <nobr>
|     <nobr>
